+++ android/src/main/java/ee/forgr/capacitor_inappbrowser/InAppBrowserPlugin.java
@@ -28,6 +28,8 @@ import com.getcapacitor.annotation.Permission;
 import com.getcapacitor.annotation.PermissionCallback;
 import java.util.ArrayList;
 import java.util.Iterator;
+import java.util.regex.Pattern;
+import java.util.regex.PatternSyntaxException;
 import org.json.JSONException;
 import org.json.JSONObject;
 
@@ -426,6 +428,18 @@ public class InAppBrowserPlugin
       Boolean.TRUE.equals(call.getBoolean("ignoreUntrustedSSLError", false))
     );
 
+    String proxyRequestsStr = call.getString("proxyRequests");
+    if (proxyRequestsStr != null) {
+      try {
+        options.setProxyRequestsPattern(Pattern.compile(proxyRequestsStr));
+      } catch (PatternSyntaxException e) {
+        Log.e(
+          "WebViewDialog",
+          String.format("Pattern '%s' is not a valid pattern", proxyRequestsStr)
+        );
+      }
+    }
+
     try {
       Options.ButtonNearDone buttonNearDone =
         Options.ButtonNearDone.generateFromPluginCall(
@@ -552,10 +566,11 @@ public class InAppBrowserPlugin
               getContext(),
               android.R.style.Theme_NoTitleBar,
               options,
-              InAppBrowserPlugin.this
+              InAppBrowserPlugin.this,
+              getBridge().getWebView()
             );
-            webViewDialog.presentWebView();
             webViewDialog.activity = InAppBrowserPlugin.this.getActivity();
+            webViewDialog.presentWebView();
           }
         }
       );
@@ -573,6 +588,7 @@ public class InAppBrowserPlugin
       call.reject("No event data provided");
       return;
     }
+
     Log.d("InAppBrowserPlugin", "Event data: " + eventData.toString());
     this.getActivity()
       .runOnUiThread(
@@ -620,6 +636,26 @@ public class InAppBrowserPlugin
             }
           }
         );
+    }
+    call.resolve();
+  }
+
+  @PluginMethod
+  public void lsuakdchgbbaHandleProxiedRequest(PluginCall call) {
+    if (webViewDialog != null) {
+      Boolean ok = call.getBoolean("ok", false);
+      String id = call.getString("id");
+      if (id == null) {
+        Log.e("InAppBrowserProxy", "CRITICAL ERROR, proxy id = null");
+        return;
+      }
+      if (Boolean.FALSE.equals(ok)) {
+        String result = call.getString("result", "");
+        webViewDialog.handleProxyResultError(result, id);
+      } else {
+        JSONObject object = call.getObject("result");
+        webViewDialog.handleProxyResultOk(object, id);
+      }
     }
     call.resolve();
   }